scripts/auto-remediate-all-tags/requirements.txt
azure-identity==1.21.0
azure-mgmt-policyinsights==1.0.0
azure-mgmt-resource==23.3.0